Refactor: Recommender 컴포넌트 빈으로 변경
