+++ src/main/java/com/newzet/api/newsletter/business/NewsletterRecommender.java
@@ -1,16 +1,19 @@
-package com.newzet.api.newsletter.util;
+package com.newzet.api.newsletter.business;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+import org.springframework.stereotype.Component;
+
 import com.newzet.api.newsletter.business.exception.NotEnoughNewslettersException;
 import com.newzet.api.newsletter.domain.Newsletter;
 import com.newzet.api.newsletter.domain.RecommendationPolicy;
 
+@Component
 public class NewsletterRecommender {
 
-	public static List<Newsletter> recommendNewsletterList(List<Newsletter> advertiseNewsletterList,
+	public List<Newsletter> recommendNewsletterList(List<Newsletter> advertiseNewsletterList,
 		List<Newsletter> userCategoryNewsletterList) {
 
 		List<Newsletter> recommendedNewsletterList = new ArrayList<>();
@@ -28,7 +31,7 @@ public class NewsletterRecommender {
 		return recommendedNewsletterList;
 	}
 
-	private static int getRemainingQuarter(int filledQuarter, int size) {
+	private int getRemainingQuarter(int filledQuarter, int size) {
 		int resQuarter = RecommendationPolicy.RECOMMENDATION_QUARTER_SIZE - filledQuarter;
 		if (size < resQuarter) {
 			throw new NotEnoughNewslettersException("추천할 뉴스레터 count 수보다 존재하는 뉴스레터 수가 적습니다.");
@@ -36,7 +39,7 @@ public class NewsletterRecommender {
 		return resQuarter;
 	}
 
-	private static List<Newsletter> createRecommendationList(List<Newsletter> candidateList,
+	private List<Newsletter> createRecommendationList(List<Newsletter> candidateList,
 		int count) {
 		List<Newsletter> randomNewsletterList = new ArrayList<>(candidateList);
 		Collections.shuffle(randomNewsletterList);
